backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
email-validator==2.1.0
bcrypt==4.1.2
requests==2.31.0